tests(cli): initial set of tests for picker
