+++ crates/but/src/command/legacy/status/tui/copy_selection_picker.rs
@@ -1,4 +1,4 @@
-use std::{borrow::Cow, collections::BTreeSet, sync::Arc};
+use std::{borrow::Cow, collections::BTreeSet};
 
 use anyhow::Context as _;
 use bstr::{BString, ByteSlice as _};
@@ -14,7 +14,6 @@ use crate::{
     },
     id::{ShortId, UncommittedHunkOrFile, WorktreeHunk},
     theme::Theme,
-    utils::DebugAsType,
 };
 
 pub fn commit_picker(
@@ -351,55 +350,4 @@ fn tree_changes_to_diff(
             diff.push_str(&line?.to_str_lossy());
             Ok(diff)
         })
-}
-
-#[derive(Clone, Debug)]
-pub struct Clipboard(DebugAsType<Arc<dyn ClipboardImpl + Send>>);
-
-impl Clipboard {
-    pub fn live() -> Self {
-        struct Live;
-
-        impl ClipboardImpl for Live {
-            fn set_text(&self, text: Cow<'_, str>) -> anyhow::Result<()> {
-                arboard::Clipboard::new()
-                    .and_then(|mut clipboard| clipboard.set_text(text))
-                    .context("failed to copy to system clipboard")?;
-                Ok(())
-            }
-        }
-
-        Self::new(Live)
-    }
-
-    #[cfg(test)]
-    pub fn test() -> (Self, Arc<std::sync::Mutex<String>>) {
-        struct Test(Arc<std::sync::Mutex<String>>);
-
-        let shared = <Arc<std::sync::Mutex<String>>>::default();
-
-        impl ClipboardImpl for Test {
-            fn set_text(&self, text: Cow<'_, str>) -> anyhow::Result<()> {
-                *self.0.lock().unwrap() = text.to_string();
-                Ok(())
-            }
-        }
-
-        (Self::new(Test(Arc::clone(&shared))), shared)
-    }
-
-    fn new<C>(clipboard_impl: C) -> Self
-    where
-        C: ClipboardImpl + Send + 'static,
-    {
-        Self(DebugAsType(Arc::new(clipboard_impl)))
-    }
-
-    pub fn set_text<'a>(&self, text: impl Into<Cow<'a, str>>) -> anyhow::Result<()> {
-        self.0.set_text(text.into())
-    }
-}
-
-trait ClipboardImpl {
-    fn set_text(&self, text: Cow<'_, str>) -> anyhow::Result<()>;
 }
